Many changes, start to count ressources for Action_Dice

+++ src/pioneers3d/types/Bank.hpp
@@ -16,7 +16,7 @@ public:
         Holz = Lehm = Weizen = Wolle = Erz = 0;
     }
 
-    uint32_t getTotal() const
+    int32_t getTotal() const
     {
         return Holz + Lehm + Weizen + Wolle + Erz;
     }
@@ -39,9 +39,9 @@ public:
         return s.str();
     }
 
-    uint32_t get( TileType type ) const
+    int32_t get( TileType type ) const
     {
-        uint32_t n = 0;
+        int32_t n = 0;
         if ( type.contains( TileType::HOLZ ) ) n += Holz;
         if ( type.contains( TileType::LEHM ) ) n += Lehm;
         if ( type.contains( TileType::WEIZEN ) ) n += Weizen;
@@ -52,28 +52,28 @@ public:
 
     void set( TileType type, int32_t value )
     {
-        if ( type.contains( TileType::HOLZ ) ) Holz = uint32_t( glm::clamp( value, 0, 16000 ) );
-        if ( type.contains( TileType::LEHM ) ) Lehm = uint32_t( glm::clamp( value, 0, 16000 ) );
-        if ( type.contains( TileType::WEIZEN ) ) Weizen = uint32_t( glm::clamp( value, 0, 16000 ) );
-        if ( type.contains( TileType::WOLLE ) ) Wolle = uint32_t( glm::clamp( value, 0, 16000 ) );
-        if ( type.contains( TileType::ERZ ) ) Erz = uint32_t( glm::clamp( value, 0, 16000 ) );
+        if ( type.contains( TileType::HOLZ ) ) Holz = glm::clamp( value, 0, 16000 );
+        if ( type.contains( TileType::LEHM ) ) Lehm = glm::clamp( value, 0, 16000 );
+        if ( type.contains( TileType::WEIZEN ) ) Weizen = glm::clamp( value, 0, 16000 );
+        if ( type.contains( TileType::WOLLE ) ) Wolle = glm::clamp( value, 0, 16000 );
+        if ( type.contains( TileType::ERZ ) ) Erz = glm::clamp( value, 0, 16000 );
     }
 
     void add( TileType type, int32_t value )
     {
-        if ( type.contains( TileType::HOLZ ) ) Holz = uint32_t( glm::clamp( int32_t( Holz ) + value, 0, 16000 ) );
-        if ( type.contains( TileType::LEHM ) ) Lehm = uint32_t( glm::clamp( int32_t( Lehm ) + value, 0, 16000 ) );
-        if ( type.contains( TileType::WEIZEN ) ) Weizen = uint32_t( glm::clamp( int32_t( Weizen ) + value, 0, 16000 ) );
-        if ( type.contains( TileType::WOLLE ) ) Wolle = uint32_t( glm::clamp( int32_t( Wolle ) + value, 0, 16000 ) );
-        if ( type.contains( TileType::ERZ ) ) Erz = uint32_t( glm::clamp( int32_t( Erz ) + value, 0, 16000 ) );
+        if ( type.contains( TileType::HOLZ ) ) Holz = glm::clamp( int32_t( Holz ) + value, 0, 16000 );
+        if ( type.contains( TileType::LEHM ) ) Lehm = glm::clamp( int32_t( Lehm ) + value, 0, 16000 );
+        if ( type.contains( TileType::WEIZEN ) ) Weizen = glm::clamp( int32_t( Weizen ) + value, 0, 16000 );
+        if ( type.contains( TileType::WOLLE ) ) Wolle = glm::clamp( int32_t( Wolle ) + value, 0, 16000 );
+        if ( type.contains( TileType::ERZ ) ) Erz = glm::clamp( int32_t( Erz ) + value, 0, 16000 );
     }
 
 public:
-    uint32_t Holz = 0;           // Anzahl Holz
-    uint32_t Lehm = 0;           // Anzahl Lehm
-    uint32_t Weizen = 0;         // Anzahl Weizen
-    uint32_t Wolle = 0;          // Anzahl Wolle
-    uint32_t Erz = 0;            // Anzahl Erz
+    int32_t Holz = 0;           // Anzahl Holz
+    int32_t Lehm = 0;           // Anzahl Lehm
+    int32_t Weizen = 0;         // Anzahl Weizen
+    int32_t Wolle = 0;          // Anzahl Wolle
+    int32_t Erz = 0;            // Anzahl Erz
     std::string Name = "Bank";
 };
